Add files via upload
PREDICTING STUDENT'S DROPOUT AND ACADEMIC SUCCESS

A major indicator of likely hitches in the higher education system is attributed to the enormous number of student dropouts in the early academic years. A careful analysis of the existing data set holds several demographic and socioeconomic factors regarding different students and their eventual successful completion or not of programs in a given higher institution. To better understand the obstacles responsible for these dropouts, the data are processed using machine learning methods such as decision trees and random forests. The results indicate that machine learning can predict with reasonable accuracy the likelihood of success for a student (i.e. graduate) however, the models suffer from an imbalance in the data class of the dataset.